[Struc] Improved project structure.
Removed views from main file and insert to respective files.
